mod: Rename `core` module to `common`
Renamed `core` module as `common` in Kyron to avoid issues when imported crates use standard `core` items which are not available in Kyron's `core`.
